refactor camera class hierarchy
