+++ pygfx/cameras/__init__.py
@@ -6,6 +6,7 @@
     Camera
     NDCCamera
     ScreenCoordsCamera
+    GenericCamera
     PerspectiveCamera
     OrthographicCamera
 
@@ -14,5 +15,6 @@
 # flake8: noqa
 
 from ._base import Camera, NDCCamera, ScreenCoordsCamera
+from ._generic import GenericCamera
 from ._perspective import PerspectiveCamera
 from ._orthographic import OrthographicCamera
